#18 refactor(addons pkg): addons loading and addons list cmd separated
