+++ .gitignore
@@ -89,11 +89,11 @@ bin/msiuninstall.log
 
 # addons
 addons/hooks
-tmp/addons
-tmp/addons.zip
 
 # apps in /k2s
 /k2s/**/*.exe
+
+tmp/**/*
 
 # temp files in /k2s
 /k2s/**/*.tmp
