+++ k2s/cmd/k2s/cmd/addons/importcmd/importcmd.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText:  © 2023 Siemens Healthcare GmbH
 // SPDX-License-Identifier:   MIT
 
-package addonimport
+package importcmd
 
 import (
 	"errors"
@@ -16,10 +16,7 @@ import (
 	p "github.com/siemens-healthineers/k2s/cmd/k2s/cmd/params"
 
 	ac "github.com/siemens-healthineers/k2s/cmd/k2s/cmd/addons/common"
-
-	acc "github.com/siemens-healthineers/k2s/cmd/k2s/cmd/addons/cmd/common"
-
-	"github.com/siemens-healthineers/k2s/cmd/k2s/addons"
+	"github.com/siemens-healthineers/k2s/internal/addons"
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/common"
 
@@ -60,14 +57,14 @@ func NewCommand() *cobra.Command {
 
 func runImport(cmd *cobra.Command, args []string) error {
 	terminalPrinter := terminal.NewTerminalPrinter()
-	allAddons, err := addons.LoadAddons()
+	allAddons, err := addons.LoadAddons(utils.InstallDir())
 	if err != nil {
 		return err
 	}
 
 	ac.LogAddons(allAddons)
 
-	if err := acc.ValidateAddonNames(allAddons, "import", terminalPrinter, args...); err != nil {
+	if err := ac.ValidateAddonNames(allAddons, "import", terminalPrinter, args...); err != nil {
 		return err
 	}
