+++ k2s/cmd/k2s/cmd/addons/status/load.go
