--- k2s/internal/terminal/defs/defs.go
@@ -1,9 +1,0 @@
-// SPDX-FileCopyrightText:  © 2023 Siemens Healthcare GmbH
-// SPDX-License-Identifier:   MIT
-
-package defs
-
-type LeveledListItem struct {
-	Level int
-	Text  string
-}
