+++ k2s/test/e2e/addons/exthttpaccess/exthttpsaccess_test.go
@@ -11,11 +11,10 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/addons/status"
 	"github.com/siemens-healthineers/k2s/test/framework"
 
 	"github.com/siemens-healthineers/k2s/test/framework/k2s"
-
-	"github.com/siemens-healthineers/k2s/cmd/k2s/addons/status"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
