+++ k2s/test/e2e/addons/kubevirt/kubevirt_test.go
@@ -14,11 +14,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/addons/status"
 	"github.com/siemens-healthineers/k2s/test/framework"
 
 	"github.com/siemens-healthineers/k2s/test/framework/k2s"
-
-	"github.com/siemens-healthineers/k2s/cmd/k2s/addons/status"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
